+++ dart/geobase/CHANGELOG.md
@@ -12,6 +12,7 @@ NOTE: Version 1.1.0 currently under development (1.1.0-dev.0).
 * [Support for GeoJSON Text Sequences](https://github.com/navibyte/geospatial/issues/217)
 * [Centroid for all geometry objects #223](https://github.com/navibyte/geospatial/issues/223)
 * [Factories to create multi geometries from geometry lists #229](https://github.com/navibyte/geospatial/issues/229)
+* [Item accessor operator and length in geometry and feature collections #230](https://github.com/navibyte/geospatial/issues/230)
 
 ## 1.0.2